chore: rename chyt api to strawberry [YTFRONT-4219]
