+++ packages/ui/src/ui/store/actions/chyt/options.ts
@@ -5,7 +5,11 @@ import {wrapApiPromiseByToaster} from '../../../utils/utils';
 import {ChytCliqueOptionsAction} from '../../reducers/chyt/options';
 import {RootState} from '../../reducers';
 import {getCluster, isDeveloper} from '../../selectors/global';
-import {ChytListAttributes, ChytListResponseItem, chytApiAction} from '../../../utils/api';
+import {
+    ChytListAttributes,
+    ChytListResponseItem,
+    chytApiAction,
+} from '../../../utils/strawberryControllerApi';
 import {CHYT_OPTIONS} from '../../../constants/chyt-page';
 import {chytCliqueLoad} from './clique';
 import {chytLoadCliqueSpeclet} from './speclet';
