+++ packages/ui/src/ui/store/reducers/chyt/options.ts
@@ -1,7 +1,7 @@
 import {ActionD, YTError} from '../../../types';
 import {mergeStateOnClusterChange} from '../utils';
 import {CHYT_OPTIONS} from '../../../constants/chyt-page';
-import {ChytDescribeOptionsType} from '../../../utils/api';
+import {ChytDescribeOptionsType} from '../../../utils/strawberryControllerApi';
 
 export type ChytCliqueOptionsState = {
     loaded: boolean;
